requirements.txt
# amqp==2.5.1
# anyjson==0.3.3
# billiard==3.6.1.0
# celery==4.3.0
# Django==2.2.3
# django-bootstrap-breadcrumbs==0.9.1
# django-braces==1.13.0
# django-countries==5.3.3
# django-grappelli==2.13.1
# django-model-utils==3.1.2
# django-registration-redux==2.6
# django-reverse-admin==2.3.0
# django-reversion==3.0.4
# django-simple-history==2.7.3
# django-suit==0.2.26
# django-user-sessions==1.6.0
# django-widget-tweaks==1.4.5
# dnspython==1.16.0
# et-xmlfile==1.0.1
# eventlet==0.25.1
# greenlet==0.4.15
# importlib-metadata==0.23
# jdcal==1.4.1
# kombu==4.6.5
# lml==0.0.9
# monotonic==1.5
# more-itertools==7.2.0
# openpyxl==2.5.14
# pillow==7.0.0
# pytz==2019.2
# six==1.12.0
# sqlparse==0.3.0
# vine==1.3.0
# zipp==0.6.0


amqp==2.5.1
anyjson==0.3.3
billiard==3.6.1.0
celery==4.3.0
Django==2.2.3
django-bootstrap-breadcrumbs==0.9.1
django-braces==1.13.0
django-countries==5.3.3
django-grappelli==2.13.1
django-model-utils==3.1.2
django-registration-redux==2.6
django-reverse-admin==2.3.0
django-reversion==3.0.4
django-simple-history==2.7.3
django-suit==0.2.26
django-user-sessions==1.6.0
django-widget-tweaks==1.4.5
dnspython==1.16.0
et-xmlfile==1.0.1
eventlet==0.25.1
greenlet==0.4.15
importlib-metadata==0.23
jdcal==1.4.1
kombu==4.6.5
lml==0.0.9
monotonic==1.5
more-itertools==7.2.0
openpyxl==2.5.14
Pillow==6.1.0
pyexcel-io==0.5.20
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.7
python-memcached==1.59
pytz==2019.2
redis==3.3.11
six==1.12.0
sqlparse==0.3.0
swapper==1.1.1
tqdm==4.35.0
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0

